backend/requirements.txt
Flask>=2.3.0,<3.1.0
python-dotenv>=1.0.0
firebase-admin>=6.0.0
Flask-Cors>=4.0.0
gunicorn>=20.1.0 # Для развертывания
# Добавьте другие зависимости, такие как pydantic, если вы его используете для моделей
# pydantic>=2.0.0